[#7]; refactor: 패키지 구조 변경
- 도메인 별로 패키지를 만들어 클래스들을 분류함.
- 전역 설정, 부모 예외, Global Exception Handler 등은 global 패키지로 이동시킴.
