--- src/main/java/com/sennyru/onboarding/domain/Member.java
@@ -1,35 +1,0 @@
-package com.sennyru.onboarding.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@RequiredArgsConstructor(staticName = "create")
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NonNull
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @NonNull
-    @Column(nullable = false)
-    private String password;
-
-    @NonNull
-    @Column(nullable = false)
-    private String username;
-}
