--- src/main/java/com/sennyru/onboarding/dto/MemberResponseDto.java
@@ -1,7 +1,0 @@
-package com.sennyru.onboarding.dto;
-
-public record MemberResponseDto(String email, String username) {
-    public static MemberResponseDto of(String email, String username) {
-        return new MemberResponseDto(email, username);
-    }
-}
