--- src/main/java/com/sennyru/onboarding/dto/SignupRequestDto.java
@@ -1,20 +1,0 @@
-package com.sennyru.onboarding.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record SignupRequestDto(
-    @NotBlank
-    @Email
-    String email,
-
-    @NotBlank
-    String password,
-
-    @NotBlank
-    String username
-) {
-    public static SignupRequestDto of(String email, String password, String username) {
-        return new SignupRequestDto(email, password, username);
-    }
-}
