--- src/main/java/com/sennyru/onboarding/exception/DomainException.java
@@ -1,14 +1,0 @@
-package com.sennyru.onboarding.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class DomainException extends RuntimeException {
-    private final HttpStatus status;
-
-    public DomainException(String message, HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-}
