--- src/main/java/com/sennyru/onboarding/exception/EmailAlreadyExistsException.java
@@ -1,9 +1,0 @@
-package com.sennyru.onboarding.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class EmailAlreadyExistsException extends DomainException {
-    public EmailAlreadyExistsException(String message) {
-        super(message, HttpStatus.CONFLICT);
-    }
-}
