--- src/main/java/com/sennyru/onboarding/exception/InvalidPasswordException.java
@@ -1,9 +1,0 @@
-package com.sennyru.onboarding.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class InvalidPasswordException extends DomainException {
-    public InvalidPasswordException(String message) {
-        super(message, HttpStatus.UNAUTHORIZED);
-    }
-}
