--- src/main/java/com/sennyru/onboarding/exception/MemberNotFoundException.java
@@ -1,9 +1,0 @@
-package com.sennyru.onboarding.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class MemberNotFoundException extends DomainException {
-    public MemberNotFoundException(String message) {
-        super(message, HttpStatus.NOT_FOUND);
-    }
-}
