--- src/main/java/com/sennyru/onboarding/exception/PostAccessDeniedException.java
@@ -1,9 +1,0 @@
-package com.sennyru.onboarding.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class PostAccessDeniedException extends DomainException {
-    public PostAccessDeniedException(String message) {
-        super(message, HttpStatus.FORBIDDEN);
-    }
-}
