--- src/main/java/com/sennyru/onboarding/exception/PostNotFoundException.java
@@ -1,9 +1,0 @@
-package com.sennyru.onboarding.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class PostNotFoundException extends DomainException {
-    public PostNotFoundException(String message) {
-        super(message, HttpStatus.NOT_FOUND);
-    }
-}
