+++ src/main/java/com/sennyru/onboarding/global/exception/DomainException.java
@@ -1,0 +1,14 @@
+package com.sennyru.onboarding.global.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class DomainException extends RuntimeException {
+    private final HttpStatus status;
+
+    public DomainException(String message, HttpStatus status) {
+        super(message);
+        this.status = status;
+    }
+}
