--- src/main/java/com/sennyru/onboarding/handler/GlobalExceptionHandler.java
@@ -1,17 +1,0 @@
-package com.sennyru.onboarding.handler;
-
-import com.sennyru.onboarding.dto.ErrorResponse;
-import com.sennyru.onboarding.exception.DomainException;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(DomainException.class)
-    public ResponseEntity<ErrorResponse> handleDomainException(DomainException ex, HttpServletRequest request) {
-        ErrorResponse errorResponse = ErrorResponse.of(ex.getStatus(), ex.getMessage(), request.getRequestURI());
-        return new ResponseEntity<>(errorResponse, ex.getStatus());
-    }
-}
