+++ src/main/java/com/sennyru/onboarding/member/controller/dto/MemberResponseDto.java
@@ -1,0 +1,7 @@
+package com.sennyru.onboarding.member.controller.dto;
+
+public record MemberResponseDto(String email, String username) {
+    public static MemberResponseDto of(String email, String username) {
+        return new MemberResponseDto(email, username);
+    }
+}
