+++ src/main/java/com/sennyru/onboarding/member/exception/EmailAlreadyExistsException.java
@@ -1,0 +1,10 @@
+package com.sennyru.onboarding.member.exception;
+
+import com.sennyru.onboarding.global.exception.DomainException;
+import org.springframework.http.HttpStatus;
+
+public class EmailAlreadyExistsException extends DomainException {
+    public EmailAlreadyExistsException(String message) {
+        super(message, HttpStatus.CONFLICT);
+    }
+}
