+++ src/main/java/com/sennyru/onboarding/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.sennyru.onboarding.member.repository;
+
+import com.sennyru.onboarding.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    boolean existsByEmail(String email);
+
+    Optional<Member> findByEmail(String email);
+}
