+++ src/main/java/com/sennyru/onboarding/post/controller/dto/PostDeleteRequestDto.java
@@ -1,0 +1,17 @@
+package com.sennyru.onboarding.post.controller.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record PostDeleteRequestDto(
+    @NotBlank
+    @Email
+    String email,
+
+    @NotBlank
+    String password
+) {
+    public static PostDeleteRequestDto of(String email, String password) {
+        return new PostDeleteRequestDto(email, password);
+    }
+}
