+++ src/main/java/com/sennyru/onboarding/post/controller/dto/PostResponseDto.java
@@ -1,0 +1,12 @@
+package com.sennyru.onboarding.post.controller.dto;
+
+public record PostResponseDto(
+    Long articleId,
+    String email,
+    String title,
+    String content
+) {
+    public static PostResponseDto of(Long articleId, String email, String title, String content) {
+        return new PostResponseDto(articleId, email, title, content);
+    }
+}
