+++ src/main/java/com/sennyru/onboarding/post/exception/PostNotFoundException.java
@@ -1,0 +1,10 @@
+package com.sennyru.onboarding.post.exception;
+
+import com.sennyru.onboarding.global.exception.DomainException;
+import org.springframework.http.HttpStatus;
+
+public class PostNotFoundException extends DomainException {
+    public PostNotFoundException(String message) {
+        super(message, HttpStatus.NOT_FOUND);
+    }
+}
