+++ src/main/java/com/sennyru/onboarding/post/repository/PostRepository.java
@@ -1,0 +1,7 @@
+package com.sennyru.onboarding.post.repository;
+
+import com.sennyru.onboarding.post.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
