--- src/main/java/com/sennyru/onboarding/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.sennyru.onboarding.repository;
-
-import com.sennyru.onboarding.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    boolean existsByEmail(String email);
-
-    Optional<Member> findByEmail(String email);
-}
