--- src/main/java/com/sennyru/onboarding/repository/PostRepository.java
@@ -1,7 +1,0 @@
-package com.sennyru.onboarding.repository;
-
-import com.sennyru.onboarding.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-}
